+++ web/index.ts
@@ -4,6 +4,7 @@ import * as typographyStyles from './styles/typography';
 
 export * from './Box/Box';
 export * from './Text/Text';
+export * from './Offset/Offset';
 export * from './types';
 
 export { backgroundStyles, foregroundStyles, typographyStyles };